+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Features
 - Display deep link to forum posts, better responsive layout #937 !1650 !1652 @ChrisOelmueller
+- New store status "permanently closed" #786 !1655 @ChrisOelmueller
 
 ## Changes
 - Removed FAQ section !1568 @chriswalg
@@ -30,7 +31,7 @@
 - Move removal of store requests to new REST endpoint !1648 @alex.simm
 - Move buddy requests to REST endpoint !1646 #847 #798 @alex.simm
 - Move registration request to the REST endpoint #819 !1632 @alex.simm
-- Use TranslatorInterface for many old translations !1637 @ChrisOelmueller
+- Use TranslatorInterface for many old translations !1637 !1655 @ChrisOelmueller
 - Prepare event invitations for REST !1627 @ChrisOelmueller
 - Some PHP linting chores and more type hints !1641 @ChrisOelmueller
 - Refactor some translations from twig to messages.de.yml #824 !1639 @chriswalg @jonathan_b